Use Specification pattern for selecting what to purge
Reasoning: will be useful to allow purge settings to be given from a
different source - the config file.

+++ simple_git_deploy/cmd_purge.py
@@ -1,20 +1,36 @@
 import argparse
 import sys
+from typing import Optional
 
-from .core import Config, Reporter, purge_deployments
 import dateparser  # type: ignore
+
+from .core import Config, PurgeSpecification, Reporter, purge_deployments
 
 
 def cmd_purge(args, config: Config, reporter: Reporter):
+    what_to_purge = _find_what_to_purge(args)
+    if what_to_purge is None:
+        reporter.error('No valid purge settings found')
+        sys.exit(1)
+
     reporter.info('Selecting git repository %s' % config.git_dir)
+
     purge_deployments(
         deploy_root=config.deploy_root,
         dry_run=args.dry_run,
-        keep_latest=args.keep_latest,
-        older_than=args.older_than,
+        what_to_purge=what_to_purge,
         git_dir=config.git_dir,
         reporter=reporter
     )
+
+
+def _find_what_to_purge(args) -> Optional[PurgeSpecification]:
+    if args.keep_latest is not None:
+        return PurgeSpecification.keep_latest(args.keep_latest)
+    elif args.older_than is not None:
+        return PurgeSpecification.discard_older_than(args.older_than)
+    else:
+        return None
 
 
 def relative_time(string):
